+++ fxmanifest.lua
@@ -53,6 +53,7 @@ client_scripts {
     'client/minigame/profiles/catalytic.lua',       -- [PR-2] furto de catalisador
     'client/minigame/profiles/serial_scratch.lua',  -- [PR-3] lixar série na bancada
     'client/minigame/profiles/bench_teardown.lua',  -- [PR-4] desmonte na marreta
+    'client/minigame/profiles/bench_catalytic.lua', -- [FIX-1.3] desmontar catalisador na bancada
     'client/minigame/profiles.lua',
     'client/minigame/fallback.lua',
     'client/minigame/core.lua',